# Allocate vals slice in parseSeqExpr with right cap