--- a/builder/static_code.go
+++ b/builder/static_code.go
@@ -963,7 +963,7 @@ func (p *parser) parseSeqExpr(seq *seqExpr) (interface{}, bool) {
 	}
 
 	// {{ end }} ==template==
-	var vals []interface{}
+	vals := make([]interface{}, 0, len(seq.exprs))
 
 	pt := p.pt
 	for _, expr := range seq.exprs {