Performance improvement for BenchmarkPigeonJSONNoMemo:

benchmark                       old ns/op     new ns/op     delta
BenchmarkPigeonJSONNoMemo-4     18091878      16412947      -9.28%